templates: use WhiteSmoke(#f5f5f5) background for verbatim env.
This change ensures verbatime/pre/code environments look sort of the
same in both HTML and PDF versions.  A light grey box with a slightly
darker border around.

Code segments now both are numbered if {.numberLines} is used for the
code fencing.

Somehow when `--listings` was enabled we lost the syntax hihglighting
in the PDF version.  Seems there's a different way of doing things in
pandoc in this mode.  Listings at least give us captioning support so,
so we can comment on code examples or ASCII art figures, which should
be better than nothing.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>